requirements.txt
pkg_resources==0.0.0
pymodbus==2.5.3
pyserial==3.5
six==1.16.0
